Add files via upload
This is the corrected version of my stock monitoring program. The only issue is that the html keys may change over time and become invalid. However I still think it shows my ability to web scrape and parse through data, even as basic as it is. I still have more to schow.